+++ Models/Dao/UserDao.cs
@@ -37,7 +37,7 @@ namespace DotnetCoreServer.Models
                     {
                         if (reader.Read())
                         {
-                            user.UserID = reader.GetInt64(0);
+                            user.UserID = reader.GetInt32(0);
                             user.FacebookID = reader.GetString(1);
                             user.FacebookName = reader.GetString(2);
                             user.FacebookPhotoURL = reader.GetString(3);
@@ -79,7 +79,7 @@ namespace DotnetCoreServer.Models
                     {
                         if (reader.Read())
                         {
-                            user.UserID = reader.GetInt64(0);
+                            user.UserID = reader.GetInt32(0);
                             user.FacebookID = reader.GetString(1);
                             user.FacebookName = reader.GetString(2);
                             user.FacebookPhotoURL = reader.GetString(3);
